+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/I18n/I18nBase.cs
@@ -70,6 +70,7 @@ namespace KRT.VRCQuestTools.I18n
         internal abstract string IMaterialConvertSettingsMainTextureBrightnessTooltip { get; }
         internal abstract string ToonLitConvertSettingsGenerateShadowFromNormalMapLabel { get; }
         internal abstract string MatCapLitConvertSettingsMatCapTextureLabel { get; }
+        internal abstract string MatCapLitConvertSettingsMatCapTextureWarning { get; }
         internal abstract string AdditionalMaterialConvertSettingsTargetMaterialLabel { get; }
         internal abstract string AdditionalMaterialConvertSettingsSelectMaterialLabel { get; }
         internal abstract string MaterialConvertTypePopupLabelToonLit { get; }